finalização da tela-consulta
