+++ client/src/components/ConsultaCard/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 export default function ConsultaCard() {
   return (
-    <div className="w-1/3 mx-auto bg-white p-12 rounded-xl shadow-lg shadow-xl">
+    <div className="w-full max-w-[600px] mx-auto bg-white p-12 rounded-xl shadow-lg shadow-xl">
       <div className="flex mb-6">
         <button className=" px-4 py-1 hover:bg-gray-50 font-bold text-2x1 font-stretch-ultra-condensed">
           <img src="/img/arrow.png" className="w-4 h-6 object-contain" />
